+++ src/background/modules/monitors/infrastructure.rs
@@ -1,3 +1,5 @@
+use std::sync::Once;
+
 use seelen_core::{
     handlers::SeelenEvent,
     system_state::{Brightness, PhysicalMonitor},
@@ -6,28 +8,32 @@ use tauri::Emitter;
 
 use crate::{
     app::get_app_handle,
-    error::Result,
-    log_error,
-    modules::monitors::MONITOR_MANAGER,
+    error::{Result, ResultLogExt},
     windows_api::{monitor::Monitor, MonitorEnumerator},
 };
 
 use super::MonitorManager;
 
-pub fn register_monitor_webview_events() -> Result<()> {
-    MONITOR_MANAGER.lock().init()?;
-    MonitorManager::subscribe(|_event| {
-        if let Ok(monitors) = get_connected_monitors() {
-            log_error!(get_app_handle().emit(SeelenEvent::SystemMonitorsChanged, monitors));
-        }
+fn get_monitor_manager() -> &'static MonitorManager {
+    static TAURI_EVENT_REGISTRATION: Once = Once::new();
+    TAURI_EVENT_REGISTRATION.call_once(|| {
+        MonitorManager::subscribe(|_event| {
+            if let Ok(monitors) = get_connected_monitors() {
+                get_app_handle()
+                    .emit(SeelenEvent::SystemMonitorsChanged, monitors)
+                    .log_error();
+            }
+        });
     });
-    Ok(())
+    MonitorManager::instance()
 }
 
 #[tauri::command(async)]
 pub fn get_connected_monitors() -> Result<Vec<PhysicalMonitor>> {
+    get_monitor_manager();
+
     let mut monitors = Vec::new();
-    for m in MonitorEnumerator::get_all_v2()? {
+    for m in MonitorEnumerator::enumerate_win32()? {
         monitors.push(m.try_into()?);
     }
     Ok(monitors)